Renaming History directory to Logs
