+++ src/pages/index.ts
@@ -5,4 +5,4 @@ export { Agent } from './Agent';
 export { Landing } from './Landing';
 export { Alerts } from './Alerts';
 export { Alert } from './Alert';
-export { Logs } from './History';
+export { Logs } from './Logs';
